image-vector/requirements.txt
# 基础 Web 框架
fastapi
uvicorn[standard]
httpx
python-multipart

# 图像处理
Pillow

# AI 与深度学习 (指定版本以保证兼容性)
transformers
torch
torchvision

# 数据处理
numpy

einops
timm